Service->Workload
I do not think we should use "workload" and "service" interchangeably. They MAY be synonymous in some deployments. But in some cases, elements of a workload will be heterogeneous and despite having the same "identity" (as a workload) will look quite different from an external perspective (e.g. the API).